fix: setting text, share button
